+++ src/common/bytestring.c
@@ -38,7 +38,7 @@
 int bsCompare(bytestring s1, bytestring s2)
 {
 	if (s1->len != s2->len) {
-		return s1->len - s2->len;
+		return (int)s1->len - (int)s2->len;
 	}
 	return memcmp(s1->val, s2->val, s1->len);
 }
